Fix updating whilst developing
The DEBUG directive wasn't working when referencing the nuget package (as at that point the DLL was built in Release mode so never did it).

Need a different way of indicating if you are in a development environment. Telligent has the developer mode app setting so use that.

If this is true will reinstall the files every time the plugins are reloaded.